refactor(tests): streamline integration test setup and remove init-db module
