+++ package.json
@@ -21,6 +21,7 @@
     "test": "jest",
     "test:ci": "jest --ci --runInBand",
     "test:coverage": "jest --coverage --ci --runInBand",
+    "test:integration": "RUN_TESTCONTAINERS=1 npx jest utils/db/__tests__/db-service.test.ts -i --runInBand --testTimeout=180000",
     "test:watch": "jest --watch",
     "prepare": "husky"
   },
