+++ utils/db/db-service.ts
@@ -1,7 +1,6 @@
-import { Pool } from "pg";
+import { Pool, PoolClient } from "pg";
 import { NostrEvent } from "../types/types";
 import { findListingBySlug } from "../url-slugs";
-import { ensureFailedRelayPublishesTable } from "./init-db";
 
 let pool: Pool | null = null;
 let tablesInitialized = false;
@@ -9,6 +8,39 @@ let initializingTables = false;
 
 // Queue for serializing cache operations
 let cacheQueue: Promise<void> = Promise.resolve();
+
+export async function ensureFailedRelayPublishesTable(
+  client: PoolClient
+): Promise<void> {
+  await client.query(`
+    CREATE TABLE IF NOT EXISTS failed_relay_publishes (
+      event_id TEXT PRIMARY KEY,
+      owner_pubkey TEXT,
+      event_data TEXT NOT NULL,
+      relays TEXT NOT NULL,
+      created_at BIGINT NOT NULL,
+      retry_count INTEGER DEFAULT 0
+    )
+  `);
+
+  await client.query(`
+    ALTER TABLE failed_relay_publishes
+    ADD COLUMN IF NOT EXISTS event_data TEXT
+  `);
+
+  await client.query(`
+    ALTER TABLE failed_relay_publishes
+    ADD COLUMN IF NOT EXISTS owner_pubkey TEXT
+  `);
+
+  // Legacy rows pre-dating the owner_pubkey column have NULL ownership and
+  // can no longer be listed, retried, cleared, or claimed by anyone, so they
+  // would otherwise sit in the table forever. Drop them once on schema setup.
+  await client.query(`
+    DELETE FROM failed_relay_publishes
+    WHERE owner_pubkey IS NULL
+  `);
+}
 
 export async function trackFailedRelayPublishRecord({
   eventId,
